requirements.txt
torch>=2.0
torchvision>=0.17
scikit-learn>=1.4
scikit-image>=0.25
albumentations>=1.4
matplotlib>=3.8
tqdm>=4.66
numpy>=1.24
pillow>=10.0
Flask>=2.0
